+++ src/housewire/__init__.py
@@ -3,7 +3,7 @@
 from housewire.i18n import about_description_for
 
 __title__ = "HouseWire"
-__version__ = "0.74.0"
+__version__ = "0.75.0"
 __author__ = "Guillermo Adrián Molina"
 __license__ = "SSPL-1.0"
 __copyright__ = "Copyright (c) 2026 Guillermo Adrián Molina"